+++ src/snowflake/cli/_plugins/project/commands.py
@@ -73,7 +73,7 @@ add_object_command_aliases(
         help_example='`list --like "my%"` lists all projects that begin with “my”'
     ),
     scope_option=scope_option(help_example="`list --in database my_db`"),
-    ommit_commands=["drop", "create", "describe"],
+    ommit_commands=["create", "describe"],
 )
 
 
@@ -154,17 +154,22 @@ def create(
 @with_project_definition()
 def add_version(
     entity_id: str = entity_argument("project"),
-    _from: Optional[str] = from_option(mutually_exclusive=["prune"]),
+    _from: Optional[str] = from_option(),
     _alias: Optional[str] = typer.Option(
         None, "--alias", help="Alias for the version.", show_default=False
     ),
     comment: Optional[str] = typer.Option(
         None, "--comment", help="Version comment.", show_default=False
     ),
-    prune: bool = PruneOption(mutually_exclusive=["_from"]),
+    prune: bool = PruneOption(default=True),
     **options,
 ):
     """Uploads local files to Snowflake and cerates a new project version."""
+    if _from is not None and prune:
+        cli_console.warning(
+            "When `--from` option is used, `--prune` option will be ignored and files from stage will be used as they are."
+        )
+        prune = False
     cli_context = get_cli_context()
     project: ProjectEntityModel = get_entity_for_operation(
         cli_context=cli_context,